Fixes for balance syncing
