+++ feature-wallet-impl/src/main/java/io/novafoundation/nova/feature_wallet_impl/data/network/blockchain/assets/transfers/statemine/StatemineAssetTransfers.kt
@@ -16,15 +16,11 @@ import io.novafoundation.nova.runtime.ext.palletNameOrDefault
 import io.novafoundation.nova.runtime.ext.requireStatemine
 import io.novafoundation.nova.runtime.multiNetwork.ChainRegistry
 import io.novafoundation.nova.runtime.multiNetwork.chain.model.Chain
-import io.novafoundation.nova.runtime.multiNetwork.chain.model.StatemineAssetId
+import io.novafoundation.nova.runtime.multiNetwork.chain.model.prepareIdForEncoding
 import io.novafoundation.nova.runtime.storage.source.StorageDataSource
 import jp.co.soramitsu.fearless_utils.runtime.AccountId
-import jp.co.soramitsu.fearless_utils.runtime.RuntimeSnapshot
-import jp.co.soramitsu.fearless_utils.runtime.definitions.types.fromHex
 import jp.co.soramitsu.fearless_utils.runtime.definitions.types.instances.AddressInstanceConstructor
 import jp.co.soramitsu.fearless_utils.runtime.extrinsic.ExtrinsicBuilder
-import jp.co.soramitsu.fearless_utils.runtime.metadata.call
-import jp.co.soramitsu.fearless_utils.runtime.metadata.module
 import jp.co.soramitsu.fearless_utils.runtime.metadata.storage
 import java.math.BigInteger
 
@@ -68,18 +64,6 @@ class StatemineAssetTransfers(
         }
 
         return assetAccount.canAcceptFunds
-    }
-
-    private fun Chain.Asset.Type.Statemine.prepareIdForEncoding(runtimeSnapshot: RuntimeSnapshot): Any? {
-        return when(val id = id) {
-            is StatemineAssetId.Number -> id.value
-            is StatemineAssetId.ScaleEncoded -> {
-                val transferCall = runtimeSnapshot.metadata.module(palletNameOrDefault()).call("transfer")
-                val assetIdType = transferCall.arguments.first().type!!
-
-                assetIdType.fromHex(runtimeSnapshot, id.scaleHex)
-            }
-        }
     }
 
     private fun ExtrinsicBuilder.statemineTransfer(
